+++ src/Backend/RubiusTestTask.Web/Controllers/MusicController.cs
@@ -24,7 +24,7 @@ namespace RubiusTestTask.Controllers
 
         // GET: api/v1/music/tracks/musician/{id}
         [HttpGet("tracks/musician/{id}")]
-        public async Task<IActionResult> GetTracksByMusician(int id)
+        public async Task<IActionResult> GetTracksByMusician(long id)
         {
             var tracks = await _musicService.GetTracksByMusicianAsync(id);
             return Ok(tracks);
@@ -32,7 +32,7 @@ namespace RubiusTestTask.Controllers
 
         // GET: api/v1/music/tracks/album/{id}
         [HttpGet("tracks/album/{id}")]
-        public async Task<IActionResult> GetTracksByAlbum(int id)
+        public async Task<IActionResult> GetTracksByAlbum(long id)
         {
             var tracks = await _musicService.GetTracksByAlbumAsync(id);
             return Ok(tracks);
@@ -40,7 +40,7 @@ namespace RubiusTestTask.Controllers
 
         // POST: api/v1/music/rate/{id}?rating={rating}
         [HttpPost("rate/{id}")]
-        public async Task<IActionResult> RateTrack(int id, [FromQuery] int rating)
+        public async Task<IActionResult> RateTrack(long id, [FromQuery] int rating)
         {
             await _musicService.RateTrackAsync(id, rating);
             return Ok();
@@ -48,7 +48,7 @@ namespace RubiusTestTask.Controllers
 
         // POST: api/v1/music/listened/{id}
         [HttpPost("listened/{id}")]
-        public async Task<IActionResult> MarkListened(int id)
+        public async Task<IActionResult> MarkListened(long id)
         {
             await _musicService.MarkTrackAsListenedAsync(id);
             return Ok();
@@ -56,7 +56,7 @@ namespace RubiusTestTask.Controllers
 
         // POST: api/v1/music/favorite/{id}
         [HttpPost("favorite/{id}")]
-        public async Task<IActionResult> AddFavorite(int id)
+        public async Task<IActionResult> AddFavorite(long id)
         {
             await _musicService.AddTrackToFavoritesAsync(id);
             return Ok();
